unit test refactoring
