+++ apps/etl-func/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "copy-avsc": "cp -R ./avro/*.avsc dist/",
-    "build": "cp -R ./avro dist && esbuild --bundle ./src/func.ts --format=esm --platform=node --target=node20 --outdir=dist --packages=external",
+    "build": "esbuild --bundle ./src/func.ts --format=esm --platform=node --target=node20 --outdir=dist --packages=external",
     "build:watch": "yarn build --watch",
     "start": "func start",
     "format": "prettier --write .",
