+++ apps/etl-func/src/adapters/__tests__/message-event.test.ts
@@ -2,10 +2,30 @@ import { aSimpleMessageEvent } from "@/__mocks__/message-event.js";
 import { Logger } from "pino";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 
-import {
-  MessageEventAdapter,
-  MessageProducerClient,
-} from "../message-event.js";
+import { messageSchema } from "../avro.js";
+import { EventHubEventProducer } from "../message-event.js";
+
+const mocks = vi.hoisted(() => ({
+  EventHubProducerClient: vi.fn().mockImplementation(() => ({
+    createBatch: () => ({
+      tryAdd: tryAddMock,
+    }),
+    sendBatch: sendBatchMock,
+  })),
+}));
+
+vi.mock("@azure/event-hubs", async (importOriginal) => {
+  const original = await importOriginal<typeof import("@azure/storage-blob")>();
+  return {
+    ...original,
+    EventHubProducerClient: mocks.EventHubProducerClient,
+  };
+});
+
+const tryAddMock = vi.fn(() => true);
+const sendBatchMock = vi.fn(() => Promise.resolve());
+
+const eventHubProducerClient = new mocks.EventHubProducerClient();
 
 const errorLogMock = vi.fn();
 
@@ -13,35 +33,44 @@ const loggerMock = {
   error: errorLogMock,
 } as unknown as Logger;
 
-const publishMessage = vi.fn();
-
-const messageProducerClient: MessageProducerClient = {
-  publishMessage,
-};
-
-const messageEventAdapter = new MessageEventAdapter(
-  messageProducerClient,
+const messageEventAdapter = new EventHubEventProducer(
+  eventHubProducerClient,
+  messageSchema,
   loggerMock,
 );
 
 describe("publishMessageEvent", () => {
   beforeEach(() => {
-    publishMessage.mockReset();
+    tryAddMock.mockClear();
+    tryAddMock.mockImplementation(() => true);
+    sendBatchMock.mockClear();
+    sendBatchMock.mockImplementation(() => Promise.resolve());
   });
-
-  test("Given a message event it should publish it", async () => {
-    publishMessage.mockResolvedValueOnce(undefined);
+  test("Given a valid message event it should resolve", async () => {
     await expect(
       messageEventAdapter.publish(aSimpleMessageEvent),
     ).resolves.toEqual(undefined);
-    expect(publishMessage).toHaveBeenCalledOnce();
+    expect(tryAddMock).toHaveBeenCalledOnce();
+    expect(sendBatchMock).toHaveBeenCalledOnce();
   });
 
-  test("should throw if publishMessage throw an error", async () => {
-    publishMessage.mockRejectedValueOnce(undefined);
+  test("Should throw an error if tryAddMock returns false", async () => {
+    tryAddMock.mockImplementation(() => false);
+    await expect(
+      messageEventAdapter.publish(aSimpleMessageEvent),
+    ).rejects.toEqual(new Error("Error while adding event to the batch"));
+    expect(tryAddMock).toHaveBeenCalledOnce();
+    expect(tryAddMock).toHaveReturnedWith(false);
+    expect(sendBatchMock).not.toHaveBeenCalledOnce();
+  });
+
+  test("Should throw an error if sendBatchMock rejects", async () => {
+    sendBatchMock.mockImplementation(() => Promise.reject());
     await expect(
       messageEventAdapter.publish(aSimpleMessageEvent),
     ).rejects.toEqual(undefined);
-    expect(publishMessage).toHaveBeenCalledOnce();
+    expect(tryAddMock).toHaveBeenCalledOnce();
+    expect(tryAddMock).toHaveReturnedWith(true);
+    expect(sendBatchMock).toHaveBeenCalledOnce();
   });
 });
